Renamed user_info to user_data for consistency. This includes renaming the database table, so existing tables should be altered accordingly.
