+++ sandpiper/common/discord.py
@@ -4,7 +4,7 @@ from typing import List, Tuple
 import discord
 from discord.ext.commands import BadArgument
 
-from sandpiper.user_info.enums import PrivacyType
+from sandpiper.user_data.enums import PrivacyType
 
 __all__ = ['date_handler', 'privacy_handler', 'find_user_in_mutual_guilds',
            'find_users_by_display_name', 'find_users_by_username']
