+++ sandpiper/sandpiper.py
@@ -48,7 +48,7 @@ class Sandpiper(commands.Bot):
                     f'You don\'t need to prefix commands here. '
                     f'Just type "{rest}".')
 
-        self.load_extension('sandpiper.user_info')
+        self.load_extension('sandpiper.user_data')
         self.load_extension('sandpiper.bios')
         self.load_extension('sandpiper.conversion')
